Changed functionailty
Users will be required to have a login account. This login account will
be linked to the account details of that person.
Added Code to verify the Username and password in the database.
Database contains a accounts table and a logindetails table which are
linked with account ID has the mutual key.